Reorganized backends tests.
