+++ tests/backends/mysql/test_features.py
@@ -4,16 +4,10 @@ from django.db import connection
 from django.test import TestCase
 
 
-class TestDatabaseFeatures(TestCase):
+@skipUnless(connection.vendor == 'mysql', 'MySQL tests')
+class TestFeatures(TestCase):
 
-    def test_nonexistent_feature(self):
-        self.assertFalse(hasattr(connection.features, 'nonexistent'))
-
-
-@skipUnless(connection.vendor == 'mysql', 'MySQL backend tests')
-class TestMySQLFeatures(TestCase):
-
-    def test_mysql_supports_transactions(self):
+    def test_supports_transactions(self):
         """
         All storage engines except MyISAM support transactions.
         """
