+++ tests/backends/postgresql/test_server_side_cursors.py
@@ -6,10 +6,10 @@ from contextlib import contextmanager
 from django.db import connection
 from django.test import TestCase
 
-from .models import Person
+from ..models import Person
 
 
-@unittest.skipUnless(connection.vendor == 'postgresql', "Test only for PostgreSQL")
+@unittest.skipUnless(connection.vendor == 'postgresql', 'PostgreSQL tests')
 class ServerSideCursorsPostgres(TestCase):
     cursor_fields = 'name, statement, is_holdable, is_binary, is_scrollable, creation_time'
     PostgresCursor = namedtuple('PostgresCursor', cursor_fields)
